Convert all ofPixel dimensions from int to size_t

+++ libs/openFrameworks/graphics/ofBitmapFont.cpp
@@ -333,13 +333,13 @@ void ofBitmapFont::init(){
 	if(pixels.isAllocated()) return;
 	pixels.allocate(16*16, 16*16, OF_PIXELS_GRAY_ALPHA); // letter size:8x14pixels, texture size:16x8letters, gl_r: 1bytes/1pixel
 	pixels.set(0);
-	for (int i = 0; i < 256; i++) {
+	for (size_t i = 0; i < 256; i++) {
 		const unsigned char * face = bmpChar_8x13_Map[i];
-		for (int j = 1; j < 15; j++){
-			for (int k = 0; k < 8; k++){
+		for (size_t j = 1; j < 15; j++){
+			for (size_t k = 0; k < 8; k++){
 				if ( ((face[15-j] << k) & (128)) > 0 ){
-					pixels[(((int)(i/16))*16*16*16+(i%16)*16 + (j-1)*16*16 + k)*2] = 255;
-					pixels[(((int)(i/16))*16*16*16+(i%16)*16 + (j-1)*16*16 + k)*2+1] = 255;
+					pixels[(((i/16))*16*16*16+(i%16)*16 + (j-1)*16*16 + k)*2] = 255;
+					pixels[(((i/16))*16*16*16+(i%16)*16 + (j-1)*16*16 + k)*2+1] = 255;
 				}
 			}
 		}
